--- a/gerrychain/partition/partition.py
+++ b/gerrychain/partition/partition.py
@@ -75,13 +75,13 @@ class Partition:
         self.parent = parent
         self.flips = flips
 
-        self.assignment = parent.assignment.copy()
-        self.assignment.update(flips)
-
         self.graph = parent.graph
         self.updaters = parent.updaters
 
         self.flows = flows_from_changes(parent.assignment, flips)
+
+        self.assignment = parent.assignment.copy()
+        self.assignment.update_flows(self.flows)
 
         if "cut_edges" in self.updaters:
             self.edge_flows = compute_edge_flows(self)