--- a/gerrychain/proposals/tree_proposals.py
+++ b/gerrychain/proposals/tree_proposals.py
@@ -39,7 +39,7 @@ def recom(
 
     """
     edge = random.choice(tuple(partition["cut_edges"]))
-    parts_to_merge = (partition.assignment[edge[0]], partition.assignment[edge[1]])
+    parts_to_merge = (partition.assignment.mapping[edge[0]], partition.assignment.mapping[edge[1]])
 
     subgraph = partition.graph.subgraph(
         partition.parts[parts_to_merge[0]] | partition.parts[parts_to_merge[1]]
@@ -65,8 +65,8 @@ def reversible_recom(partition, pop_col, pop_target, epsilon,
     def dist_pair_edges(part, a, b):
         return set(
             e for e in part.graph.edges
-            if ((part.assignment[e[0]] == a and part.assignment[e[1]] == b) or
-                (part.assignment[e[0]] == b and part.assignment[e[1]] == a))
+            if ((part.assignment.mapping[e[0]] == a and part.assignment.mapping[e[1]] == b) or
+                (part.assignment.mapping[e[0]] == b and part.assignment.mapping[e[1]] == a))
         )
 
     def bounded_balance_edge_fn(*args, **kwargs):
@@ -95,7 +95,7 @@ def reversible_recom(partition, pop_col, pop_target, epsilon,
         return partition    # self-loop: no adjacency
 
     edge = random.choice(list(pair_edges))
-    parts_to_merge = (partition.assignment[edge[0]], partition.assignment[edge[1]])
+    parts_to_merge = (partition.assignment.mapping[edge[0]], partition.assignment.mapping[edge[1]])
     subgraph = partition.graph.subgraph(
         partition.parts[parts_to_merge[0]] | partition.parts[parts_to_merge[1]]
     )