--- a/gerrychain/tree.py
+++ b/gerrychain/tree.py
@@ -201,7 +201,7 @@ def bipartition_tree(
         tree is not provided
     :param choice: :func:`random.choice`. Can be substituted for testing.
     """
-    populations = {node: graph.nodes[node][pop_col] for node in graph.node_indicies}
+    populations = {node: graph.lookup(node, pop_col) for node in graph.node_indicies}
 
     possible_cuts = []
     if spanning_tree is None:
@@ -231,7 +231,7 @@ def _bipartition_tree_random_all(
     choice=random.choice,
 ):
     """Randomly bipartitions a graph and returns all cuts."""
-    populations = {node: graph.nodes[node][pop_col] for node in graph.node_indicies}
+    populations = {node: graph.lookup(node, pop_col) for node in graph.node_indicies}
 
     possible_cuts = []
     if spanning_tree is None:
@@ -342,7 +342,7 @@ def recursive_tree_part(
         part_pop = 0
         for node in nodes:
             flips[node] = part
-            part_pop += graph.nodes[node][pop_col]
+            part_pop += graph.lookup(node, pop_col)
         debt += part_pop - pop_target
         remaining_nodes -= nodes
 
@@ -385,7 +385,7 @@ def get_seed_chunks(
 
     chunk_pop = 0
     for node in graph.node_indicies:
-        chunk_pop += graph.nodes[node][pop_col]
+        chunk_pop += graph.lookup(node, pop_col)
 
     while True:
         epsilon = abs(epsilon)
@@ -425,7 +425,7 @@ def get_seed_chunks(
 
         part_pop = 0
         for node in remaining_nodes:
-            part_pop += graph.nodes[node][pop_col]
+            part_pop += graph.lookup(node, pop_col)
         part_pop_as_dist = part_pop / num_chunks_left
         fake_epsilon = epsilon
         if num_chunks_left != 1: