--- a/gerrychain/updaters/county_splits.py
+++ b/gerrychain/updaters/county_splits.py
@@ -35,7 +35,7 @@ def compute_county_splits(partition, county_field, partition_field):
     if not partition.parent:
         county_dict = dict()
 
-        for node in partition.graph:
+        for node in partition.graph.nodes:
             county = partition.graph.nodes[node][county_field]
             if county in county_dict:
                 split, nodes, seen = county_dict[county]