--- a/gerrychain/updaters/tally.py
+++ b/gerrychain/updaters/tally.py
@@ -102,14 +102,13 @@ class Tally:
         :param partition: :class:`Partition` class.
         """
         parent = partition.parent
-        flips = partition.flips
 
         old_tally = parent[self.alias]
         new_tally = dict(old_tally)
 
         graph = partition.graph
 
-        for part, flow in flows_from_changes(parent.assignment, flips).items():
+        for part, flow in flows_from_changes(parent, partition).items():
             out_flow = compute_out_flow(graph, self.fields, flow)
             in_flow = compute_in_flow(graph, self.fields, flow)
             new_tally[part] = old_tally[part] - out_flow + in_flow