This is a work-in-progress, draft PR with some performance enhancements. In particular,
- `__slots__` is enabled for a few frequently-created classes (`Assignment` and `SubgraphView`), resulting in minor performance improvements
- caching for iterating over the neighbors of all flips is implemented to improve flow and cut edge calculation runtimes
- caching for iterating over the neighbors of a node in a graph is implemented (since `networkx`'s neighbors function is quite expensive and a partition's graph will always remain static)
- `Graph.nodes` and `Graph.edges` is cached

Combined, these speedups improve *both* standard GerryChain run performance and replay performance. In particular, the time it takes to run 100 steps on PA goes from 29.8 seconds to 26 seconds (~13% decrease) and the time spent on each loop when replaying goes from 610 ms (based on #363) to 469 ms (~23% decrease).

Some caveats:
- I used `functools.cache`, which is a Python 3.9 feature. We should switch to `functools.lru_cache` or something with better compatibility with older versions of Python before we merge this in.
- For caching reasons, GerryChain Partitions will no longer accept a `networkx` graph (to convert a `networkx` graph to a `gerrychain.Graph` object, a new method `Graph.from_networkx` is introduced in this PR). I had to rewrite the tests accordingly.