@InnovativeInventor 2022-01-02
Note that the behavior of `Graph.nodes` and `Graph.edges` will not be backwards-compatible. It may be better to rename the cached versions to `Graph._cached_nodes` and `Graph._cached_edges`.