@InnovativeInventor 2022-01-04
Some more benchmark stats with 11906d0 (replay perf on 1000-step chains, PA congressional):
- this PR (#372): 220 it/s
- Parker's PR (#363): 170 it/s
- main branch: 72 it/s

The `FrozenGraph` class also opens the door to using faster read-only data structures (generated when the class is instantiated) behind the scenes without breaking backwards-compatibility.