@InnovativeInventor 2022-01-05
Updated stats (wrote some better benchmarks):
```
- 247 it/s (my perf tuning + Parker's slots)
- 188 it/s (Parker's slots)
- 82 it/s (main branch)
```