@InnovativeInventor 2022-01-05
I realized that I was profiling #363 incorrectly -- the performance is now as follows:
- 353 it/s (my perf tuning + Parker's slots)
- 194 it/s (Parker's slots)
- 189 it/s (main branch)

My bad -- looks like `__slots__` doesn't help that much after all.