@InnovativeInventor 2022-01-11
Memory leak!: https://github.com/mggg/GerryChain/blob/c55f38260b196cc1a20c1d0c28b6eb1dbbb0b35c/gerrychain/partition/partition.py#L121