@InnovativeInventor 2022-01-19
Reminder to self: `Partition.crosses_parts` caching increases performance (but the previous impl introduced a memory leak).